WS6: promote WorldRenderBridge into the dartpy package
The ECS simulation World has no direct C++ viewer path; the supported way to
render it is WorldRenderBridge, which mirrors an sx.World into a parallel
classic render World that the Filament viewer draws. It lived as demo-local
infrastructure (python/examples/demos/_world_bridge.py); promote it to a
first-class dartpy helper so dartpy.simulation World users (after the upcoming
name-swap) have a supported GUI path.

- Move _world_bridge.py -> python/dartpy/_world_render_bridge.py (git mv; the
  module's only dartpy dependency, 'import dartpy as dart', is used lazily so it
  is safe under the package self-import).
- Attach it as dartpy.gui.WorldRenderBridge via an _install_world_render_bridge()
  hook in __init__.py (mirrors _install_simulation_experimental_diff; skipped
  when the GUI build is absent).
- Install the new pure-Python file (build-tree copy + both wheel/site install
  blocks).
- Leave a thin re-export shim at python/examples/demos/_world_bridge.py so the
  ~34 demo scenes keep importing WorldRenderBridge; new code should use
  dartpy.gui.WorldRenderBridge.

Verified: dartpy builds; dartpy.gui.WorldRenderBridge resolves and the shim
re-exports the same object; 24 bridge-dependent demo/loop-closure tests pass;
check-lint-py + gersemi clean.
