+++ python/dartpy/CMakeLists.txt
@@ -418,6 +418,7 @@ foreach(
   _naming.py
   _layout.py
   _simulation_experimental_diff.py
+  _world_render_bridge.py
 )
   configure_file(
     ${CMAKE_CURRENT_SOURCE_DIR}/${_py_file}
@@ -470,6 +471,7 @@ if(DEFINED SKBUILD_PLATLIB_DIR)
       ${CMAKE_CURRENT_SOURCE_DIR}/_naming.py
       ${CMAKE_CURRENT_SOURCE_DIR}/_layout.py
       ${CMAKE_CURRENT_SOURCE_DIR}/_simulation_experimental_diff.py
+      ${CMAKE_CURRENT_SOURCE_DIR}/_world_render_bridge.py
     DESTINATION "${SKBUILD_PLATLIB_DIR}/dartpy"
   )
 elseif(DEFINED PYTHON_SITE_PACKAGES)
@@ -480,6 +482,7 @@ elseif(DEFINED PYTHON_SITE_PACKAGES)
       ${CMAKE_CURRENT_SOURCE_DIR}/_naming.py
       ${CMAKE_CURRENT_SOURCE_DIR}/_layout.py
       ${CMAKE_CURRENT_SOURCE_DIR}/_simulation_experimental_diff.py
+      ${CMAKE_CURRENT_SOURCE_DIR}/_world_render_bridge.py
     DESTINATION "${PYTHON_SITE_PACKAGES}/dartpy"
   )
 endif()
